==> spain_barcelona/030-training_models/2.EDA.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

df = pd.read_csv('1.filtered_df.csv')
df.index = df['Unnamed: 0']
df.drop('Unnamed: 0', axis = 1, inplace = True)
df

df.describe()

# initialize the visualization matrix
viz_matrix = np.zeros([int(df.row.max()), int(df.column.max())])

values = df.iloc[:, 6:].sum(axis = 1).to_list()
rows = df.row.to_list()
columns = df.column.to_list()

for value, row, column in zip(values, rows, columns):
    viz_matrix[int(row) - 1, int(column) - 1] = value

# normalization
viz_matrix_normalized = (viz_matrix - viz_matrix.mean()) /viz_matrix.std()

# remove outliers
normalization_intensity = 0.01
low, high = np.percentile(viz_matrix_normalized, normalization_intensity * 100), np.percentile(viz_matrix_normalized, (1 - normalization_intensity) * 100)

viz_matrix_normalized = np.where(viz_matrix_normalized < low, low, viz_matrix_normalized)
viz_matrix_normalized = np.where(viz_matrix_normalized > high, high, viz_matrix_normalized)

# plotting
fig, ax = plt.subplots(figsize=(20,10))

cmap = sns.cubehelix_palette(light = 1,
                             dark = 0,
                             hue = 0,
                             as_cmap=True)

sns.heatmap(np.rot90(viz_matrix_normalized, k = 3), cmap = cmap)
