==> spain_barcelona/030-training_models/7.generating_prediction_map.ipynb <==
import pandas as pd
import numpy as np
import sklearn
import matplotlib as plt
from pycaret.regression import *

df = pd.read_csv('5.extended_df.csv')
df.index = df['Unnamed: 0']
df.drop(['Unnamed: 0', 'row', 'column', 'lat_min', 'lat_max', 'lon_min', 'lon_max'], axis = 1, inplace = True)
df.drop(list(df.filter(regex = 'tourism_')),axis = 1, inplace = True)
df

df.describe()

splits = 3
splitted_df_list = np.array_split(df, splits)

prediction_df_list = []

for splitted_df in splitted_df_list:
    training_df = df.drop(splitted_df.index)
    prediction_df = splitted_df
    
    exp = setup(training_df, target = 'tourism')
    best = compare_models()
    predictions = predict_model(best, data = prediction_df)
    
    predictions = predictions[['prediction_label']]
    predictions.index = prediction_df.index
    prediction_df_list.append(predictions)

predictions_concat = pd.concat(prediction_df_list)

export_df = pd.read_csv('3.final_map_converted.csv')
export_df.drop('Unnamed: 0', axis = 1, inplace = True)
export_df.set_index('field_1', inplace = True)
export_df

export_df['predicted_tourism'] = predictions_concat.prediction_label
export_df

export_df.to_csv('7.final_predictions.csv')
